files moved add listeners returns null
