+++ scripts/dashboardInit.js
@@ -3,7 +3,7 @@
  * Purpose is to handle initializing the structure for the dashboard
  */
 // get page elements
-const chatWidgetInit = require
+const chatWidgetInit = require("./chatWidgetInit");
 
 const welcome = document.querySelector(".welcome");
 const login = document.querySelector(".login");
@@ -18,7 +18,7 @@ const dashboardInit = function(user) {
     dashboard.style.display = "block";
 
     // const DB = getDatabase()
-    chatWidgetInit()
+    chatWidgetInit;
 }
 
 module.exports = dashboardInit;
